backend/requirements.txt
Django==4.2.20
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.1.0
stripe==7.8.1
sendgrid==6.10.0
django-sendgrid-v5==1.2.3
python-dotenv==1.0.0
psycopg2-binary==2.9.9
supabase-py==2.3.4
postgrest-py==0.15.1 